MEGA REFACTOR -> VERSION 3.0.0
Basically recreated everything from zero, with the idea of api. The only difference is that api is useless and will be removed, the new one is called 'core'.
The core package is made of interfaces.
The implementations need to be created in a different package.
That's all, pretty straight forward.